front_end/requirements_1.txt
python-dotenv
opencv-python
mediapipe==0.10.10
numpy==1.23.5
tensorflow==2.12.0
scikit-learn==1.2.2
protobuf==3.20.3
streamlit
